refactor: move modules for bot mechanism to separated file
