+++ lib/game.h
@@ -1,7 +1,6 @@
 #if !defined(game_H)
 #define game_H
-#include "../include/boolean.h"
-#include "board.h"
+#include "bot.h"
 
 #define PLAYER1 1
 #define PLAYER2 2
@@ -9,19 +8,6 @@
 
 #define PLAYER1SYMBOL "X"
 #define PLAYER2SYMBOL "O" // Also used for bot/computer
-
-typedef struct
-{
-  int row, col;
-} Move;
-
-bool isMovesLeft(Board **tempBoard);
-
-int evaluate(Board **tempBoard);
-
-int minimax(Board **tempBoard, int depth, int maxDepth, bool isMax);
-
-Move findBestMove(Board **tempBoard);
 
 bool rowCrossed(Board **board);
 
@@ -45,13 +31,6 @@ void inputUsername(char *username);
  * Procedure for input/fill the game board.
  */
 void inputMove(int *move, char *playername);
-
-/**
- * Procedure for randomize bot's movement.
- */
-void randomizeMove(int moves[]);
-
-void playPvC(Board **board, Board **hintBoard, char *player, int whoseTurn, int botDiff);
 
 void playPvP(Board **board, Board **hintBoard, char *player1, char *player2, int whoseTurn);
